package.json: remove ansi-regex 5.0.1 override and update as possible
With the override, npm was somehow incorrectly NOT giving
node_modules/ansi-align/node_modules/strip-ansi a copy of 5.0.1
even though strip-ansi's requires ^5.0.1.
Instead it got 6.2.2, which caused an exception when used
(since 6.2.2 does not have the 5.0.1 API).

With this change, as seen in package-lock.json, now there is
a copy of 5.0.1 in node_modules/ansi-align/node_modules/ansi-regex,
which will be used for ansi-align's use of strip-ansi and avoids
the exception.

The exception happens in an exit-time print from inside update-notifier
if it cannot read something in the user's home directory, which happens
if the macOS sandbox is enabled with the default rules
update-notifier is calling boxen, which calls ansi-align, which calls strip-ansi,
which calls ansi-regex.

Fixes #11812.